% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup_functions.R
\name{calculate_pricing_error}
\alias{calculate_pricing_error}
\title{Calculate Pricing Error}
\usage{
calculate_pricing_error(mar_price, mod_price, error_type)
}
\arguments{
\item{mar_price}{Market price of the contract.}

\item{mod_price}{Model price of the contract.}

\item{error_type}{The pricing error type. It will be automatically calculated. Choose from (Absolute) Relative Pricing Error (denoted with RPE or ARPE), squared error (denoted with SE) or absolute pricing error (APE).}
}
\description{
Calculate Pricing Error
}

