% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_functions.R
\name{dm_learn}
\alias{dm_learn}
\title{DM Learn}
\usage{
dm_learn(raw_data, model_name = "svm", CallPut = "call", randseed = 0,
  export_plots = FALSE, ...)
}
\arguments{
\item{raw_data}{The option data set given in the format of \link{uslfin_ds_1}.}

\item{model_name}{The name of the data mining model to use. Currently there are three options: Support Vector Machine (\code{"svm"}) of \code{e1071} package, Decision Tree (\code{"dt"}) of \code{rpart} package and Conditional Inference Tree \code{("cit")} of the \code{partykit} package.}

\item{CallPut}{It denotes whether to use the call or put options.}

\item{randseed}{To set the randomness seed to a known value. Good for reproducibility.}

\item{...}{Parameters to pass to data mining algorithm functions. (See individual functions for further explanations.)}
}
\description{
Implements different data mining algorithms (SVM, CIT and DT).
}

