% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_functions.R
\name{kmeans_learn}
\alias{kmeans_learn}
\title{K-Means Learn}
\usage{
kmeans_learn(raw_data, CallPut = "call", randseed = 0,
  moneyness_interval = c(0.9, 1.1), maturity_interval = c(4, 252),
  n_cluster = 0, export_plots = FALSE)
}
\arguments{
\item{raw_data}{The option data set given in the format of \link{uslfin_ds_1}.}

\item{CallPut}{It denotes whether to use the call or put options.}

\item{randseed}{To set the randomness seed to a known value. Good for reproducibility.}

\item{moneyness_interval}{Minimum and maximum of moneyness values. Required for rescaling.}

\item{maturity_interval}{Minimum and maximum of maturity values. Required for rescaling. In trading days (one year \= 252 days)}

\item{n_cluster}{Number of clusters. If \code{0}, then number of clusers is determined by the contract set size. Exact calculation is \code{min(round(nrow(training_matrix)/100),200)}}
}
\description{
Implements a specialized version of K-Means algorithm on the data set. When creating the clusters (in-sample) the function uses pricing_error, moneyness and maturity. But when predicting, it uses only moneyness and maturity covariates. All covariates are scaled between 0-100.
}

