% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_functions.R
\name{plot_export_manual}
\alias{plot_export_manual}
\title{Plot Export (Manual Method)}
\usage{
plot_export_manual(the_df, moneyness_breaks, maturity_breaks)
}
\description{
Explanatory plots for manual method.
}

