% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{run_full_experiment}
\alias{run_full_experiment}
\title{Run Full Experiment}
\usage{
run_full_experiment(data_set = "learnfin_ds_1", error_type = "ARPE",
  method = "kmeans", path_name = getwd(),
  file_name = "input_parameters.xlsx", randseed = 0,
  verbal_feedback = TRUE, export_raw_results = FALSE,
  export_folder = "results", summary_to_excel = FALSE)
}
\arguments{
\item{data_set}{Name of the data frame to be used in quotes (not as an object).}

\item{error_type}{Pricing error type. It can be ARPE, RPE, SE or APE.}

\item{method}{Data mining method. It can be svm, cit, dt, kmeans or manual.}

\item{path_name}{Path of the input parameter excel file. (Default: working directory \code{getwd()})}

\item{file_name}{Name of the input parameter excel file.}

\item{randseed}{Randomness seed for reproducibility.}

\item{verbal_feedback}{Do you want to be informed during the progress? (Useful for long runs.)}

\item{export_raw_results}{Should resulting data frames be exported as RData files?}

\item{export_folder}{Result raw file export folder.}

\item{summary_to_excel}{Should prediction error summaries automatically be exported to excel files?}
}
\description{
This function helps to automate a series of experiments with predetermined parameters sets with the help of an xlsx file. An example file is given with the package on the GitHub page. Experiment results can be exported to RData and xlsx files with convenient file names.
}

